Move test program to package db
